# openapi/v1.yaml
components:
  responses:
    Forbidden:
      content: &id001
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      description: Authenticated but not allowed (role, scope, tenant, quota).
    InternalError:
      content: *id001
      description: Unexpected error (a fixed body; details stay in the server log).
    Invalid:
      content: *id001
      description: The request is not valid.
    NotFound:
      content: *id001
      description: Not found (also for resources of another tenant).
    OtherError:
      content: *id001
      description: Any other error status, e.g. 409 conflict, 410 gone, 413 too large.
    TooManyRequests:
      content: *id001
      description: Rate limit or quota exceeded.
      headers:
        Retry-After:
          schema:
            minimum: 1
            type: integer
    Unauthorized:
      content: *id001
      description: Missing, invalid, expired or revoked credential.
      headers:
        WWW-Authenticate:
          schema:
            type: string
  schemas:
    AcceptedOut:
      properties:
        status:
          const: pending-confirmation
          default: pending-confirmation
          title: Status
          type: string
      title: AcceptedOut
      type: object
...
    ConfirmIn:
      additionalProperties: false
      properties:
        token:
          maxLength: 200
          minLength: 20
          title: Token
          type: string
      required:
      - token
      title: ConfirmIn
      type: object
    ConfirmedOut:
      properties:
        status:
          const: confirmed
          default: confirmed
          title: Status
          type: string
      title: ConfirmedOut
      type: object
...
    EarlyAccessIn:
      additionalProperties: false
      properties:
        email:
          maxLength: 254
          minLength: 3
          title: Email
          type: string
        organisation:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          title: Organisation
        role:
          enum:
          - researcher
          - engineer
          - clinician
          - founder
          - student
          - other
          title: Role
          type: string
        website:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          description: 'Honeypot: leave empty (hidden from people).'
          title: Website
      required:
      - email
      - role
      title: EarlyAccessIn
      type: object
...
    PipelineVersionOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        created_by:
          title: Created By
          type: string
        id:
          description: pv:sha256:<64 hex>, the content address
          title: Id
          type: string
        name:
          title: Name
          type: string
        ref:
          description: name@version
          title: Ref
          type: string
        spec:
          additionalProperties: true
          description: The published PipelineVersion document.
          title: Spec
          type: object
        version:
          title: Version
          type: string
      required:
      - id
      - name
      - version
      - ref
      - spec
      - created_by
      - created_at
      title: PipelineVersionOut
      type: object
...
    Problem:
      additionalProperties: true
      description: RFC 9457 problem details. Never contains stack traces, SQL, host names or credentials
        (SEC-079).
      properties:
        detail:
          type: string
        errors:
          description: 'Validation errors (422): location and message, never the input.'
          items:
            properties:
              loc:
                items:
                  type: string
                type: array
              msg:
                type: string
            type: object
          type: array
        instance:
          description: The request path.
          type: string
        request_id:
          description: Also in the X-Request-Id header.
          type: string
        status:
          maximum: 599
          minimum: 400
          type: integer
        title:
          type: string
        type:
          description: urn:nf:problem:<kind> or about:blank
          type: string
      required:
      - type
      - title
      - status
      type: object
...
  securitySchemes:
    apiKey:
      bearerFormat: nfb_live_<public id>_<secret>
      description: Scoped, expiring API key (SEC-014). Revocation is effective at once for new requests
        and within 60 s for open streams (SEC-017).
      scheme: bearer
      type: http
    oidc:
      description: 'User access token from the identity provider (Authorization: Bearer).'
      openIdConnectUrl: https://id.example.invalid/.well-known/openid-configuration
      type: openIdConnect
info:
  description: 'Platform API v1. Data flows device -> SDK -> platform only: no operation, event or field
    sends anything to acquisition hardware (SEC-090). Errors are RFC 9457 application/problem+json. Only
    additive changes within v1; deprecations carry Deprecation and Sunset headers.'
  title: Platform API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/pipelines:
    post:
      operationId: publishPipeline
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              title: Body
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineVersionOut'
          description: Identical re-publish
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineVersionOut'
          description: Published
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Publish Pipeline
      x-nf-action: pipeline:create
      x-nf-api-key-scope: metadata:write
  /v1/pipelines/{pipeline_ref}:
    get:
      operationId: getPipeline
      parameters:
      - in: path
        name: pipeline_ref
        required: true
        schema:
          title: Pipeline Ref
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineVersionOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Get Pipeline
      x-nf-action: pipeline:read
      x-nf-api-key-scope: metadata:read
...
  /v1/public/early-access:
    options:
      operationId: earlyAccessPreflight
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: No Origin header.
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Origin not allowed (CORS).
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security: []
      summary: Early Access Preflight
      x-nf-action: public
      x-nf-status: disabled
    post:
      operationId: joinEarlyAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EarlyAccessIn'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOut'
          description: Successful Response
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Origin not allowed (CORS).
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security: []
      summary: Join Early Access
      x-nf-action: public
      x-nf-status: disabled
  /v1/public/early-access/confirm:
    options:
      operationId: earlyAccessConfirmPreflight
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: No Origin header.
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Origin not allowed (CORS).
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security: []
      summary: Early Access Confirm Preflight
      x-nf-action: public
      x-nf-status: disabled
    post:
      operationId: confirmEarlyAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmedOut'
          description: Successful Response
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Origin not allowed (CORS).
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unknown or expired token.
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security: []
      summary: Confirm Early Access
      x-nf-action: public
      x-nf-status: disabled
...
servers:
- url: https://api.example.invalid
...